+++ dashboard/src/main/home/app-dashboard/new-app-flow/tabs/utils.ts
@@ -1,0 +1,4 @@
+export const DATABASE_HEIGHT_ENABLED = 374;
+export const DATABASE_HEIGHT_DISABLED = 119;
+export const RESOURCE_HEIGHT_WITHOUT_AUTOSCALING = 373;
+export const RESOURCE_HEIGHT_WITH_AUTOSCALING = 713;
